+++ CHANGELOG.md
@@ -11,12 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allow to wait for `SettingsHolder` to be initialized
 - Introduce `MatrixMessengerWorker` and `MatrixMultiMessengerWorker`
+- Autofill (password manager) support
 
 ### Changed
 
 - Debounce unread message counter in timeline to avoid flickering
 - Focus message input text field when starting a reply
 - Increase size of avatar and change order of settings in profile view
+- Use latest edit of an event for reader and isRead calculations
 
 ### Deprecated
 
@@ -25,7 +27,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix Database can't be read on start up
+- Update Trixnity to 4.13.1
 - Delete reply on message delete
+- only create new viewModel for timeline elements when the event content changes
+- show correct error message colors on image details view
+- TextFields behave according to their role
 
 ### Security
 
@@ -59,7 +65,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix possible crash when entering rooms with large messages
 - Remove broken collapsed message bubbles
 - Empty reactions or RedactedEventContent do not block reaction computation
-- TextFields behave according to their role
 
 ### Security
 
@@ -71,7 +76,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduced String.forEachGrapheme extension for iterating perceived chars
 - Possibility to change MatrixClientServerApiClientFactory via the DI
 - Introduce TextFieldViewModel for cleaner text input synchronization
-- Autofill (password manager) support
 
 ### Changed